requirements.txt
django
djangorestframework
pypsst